at-tag main properties must now be on single line.
Only the comment may span several lines.

This means we can just say:

  /**
   * @cfg
   * True to enable this component.
   */
  enabled: false,

Previously this meant that we declared cfg with name "True",
but now the name cannot come after newline, so we can leave the
name out of the doc-comment and infer it from the source.

Parser enhancements to be able to infer @cfg and @event name from code.

DocComment enhancements to be able to set the name of at-tags other than
the default.