Class for creating namespace-tree-structure.
JsDuck::Tree creates from doc-object list a tree structure consisting
of Ruby Hashes and Arrays.  Another method in jsduck.rb converts this
to JSON and outputs.

For all this added a unit test, but with it's single test it's mostly
just a sanity check.

The previous debug output was removed to separate method.