Different icons for Ext components.
Implemented Class.inherits_from? through which we check if
class inherits from Ext.Component (or if needed, from any other
class).

Unlike in ext-doc we mark the Ext.Component itself also with component
icon.  It always had bothered me that the base component itself was
not easily identifiable as component.