Handling multiple definitions of one class.
Reverse-engineered the behaviour of ext-doc when merging multiple
doc-blocks for the same class.  Not 100% the same, but mostly.

Now finally we have support for most of the stuff that ext-doc does
on parsing-side.  The output-side needs quite a bit more effort,
but parsing and analyzing of input seems almost complete now.