Replaced Maruku with RDiscount.
RDiscount is way faster and it doesn't throw errors for malformed
HTML like Maruku - so we can treat all the comments as Markdown,
plus we can mix HTML and Markdown.

Actually RDiscount doesn't throw any errors at all, whatever we
feed it.  If it doesn't recognize text as correct Markdown, it
doesn't complain - it just leaves it as is, which is perfect for
us.

There are some minor issues still with it:

- <pre><code>-s getting nested when parsing Ext source.
- Even one-liners are wrapped inside <p>...</p>.

But these are pretty easily solvable.