Implemented @alternateClassNames tag.
The name is now in plural (although the singular @tag works too,
but inside JSDuck the name is now "alternateClassNames" and is also
exported with that name).

Overall it behaves like @mixins tag, which too can contain multiple
items.

Correctly implemented merging of mixins and alternateClassNames when
merging class definitions.

Alternate classnames are also shown in JSDuck generated docs.