Created SourceFile class.
All parsing and merging now happens inside that class.
Consequently the Agggregator is now a lot simpler, as
it takes for input just instance of SourceFile.

SourceFile instances contain all the input: filename, source, and
parsed doc-objects.

The output of HTML source files now happens after all the
parsing and aggregating, similarly to all other output.
For now the links to source code are still line-nr based,
but this will now be easy to change.