Loading of class docs over JsonP.
Instead of pure JSON-export JSDuck now exports a .js file for each class
with the following syntax:

    Ext.data.JsonP.Ext_grid_Panel({...json-data-here...});

On JavaScript side we specify that the name of callback function should
be "Ext_grid_Panel" - this way we can use the JsonP deterministically,
knowing at JSDuck export time which names the callback functions will
have.

Note that we are passing callbackName option to Ext.data.JsonP.request,
which is not present in ExtJS-4.0.0 - I just added it to ExtJS SDK and
sent a pull request.