Ignore private classes.
Make Class#lookup return nil if class not found.

From now on, running JSDuck on ExtJS 4.0.1 will generate a bunch of
warnings as some classes that should not be private are marked as
such in ExtJS source code.

Use the --private-classes option to turn showing of private classes
back on again.