Store all class members inside :members sub-hash.
Instead of putting them directly to class hash, we create a sub-hash
for members.  This way we don't have to list all the member types
in every place where we want to loop over all the types - we just
loop over the :members hash.

Additionally it will make easier in the future to add a separate
group of static members.