Hide favorites panel when no favorites.
No more add the default top10 favorites, leave the favorites empty if
it's empty.  Automatically show the favorites panel when first item
marked as favorite.

Moved the tabpanel with favorites to separate class.

The Docs.Favorites is now observable, firering "add" and "remove" events
which are listened both by the tree and the grid.  The Docs.Favorites no
more knows anything about the tree.