Read guides from JSON file located in SDK.
--guides option now must point to a JSON file.
--guides-order option removed.

The file is read in, parsed, and the guides listed in it are
copied to output_dir/guides + the usual Markdown conversion.

Additionally the list of guides is written to tree.js as Docs.guides.
So no separate .js file for guides is needed.

The guides HTML is no more written inside index.html