New behavior for --title option.
The --title now specifies both the <title> and the contents of the header
at the very top of docs app.  The new default is "Sencha Docs - Ext JS".

The title is no more used in classes index page.  This has now static heading
"API Documentation".  Also the guides, videos, examples index pages have
headings without "Ext JS" in them - so the headings will work also for
Touch and whatever else.

Finally the title is also used in print version.  For this did some
refactoring with writing the template files in app.rb.