Refactor :compress task in Rakefile.
This task is now run *before* other tasks.  So we first compress
the contents of template/ dir into template-min/ and from then on
use the template-min/ dir as our template.

To perform sencha build, I added build-js.html file that is like
template.html but contains only bare minumum of JavaScript links
that we need to build into one file.

The purpose of all this is to run the :compress before building
gem for release version of JSDuck.  For this reason we also don't
concatenate data.js with app.js, as the jsduck gem wants to generate
its own data.js.

The default value for --template option is now "template-min/",
only in development related tasks :sdk and :touch do we replace
it with "template/".