Support for ptype and ftype in addition to xtype.
The xtypes property now basically reflects the alias property in
Ext.define class definition. It's a hash with structure like:

    xtype: {
        widget: ["text", "textfield"],
        feature: ["blah"]
    }

When rendering the class header "widget" is mapped to "xtype",
"feature" to "ftype", "plugin" to "ptype".

Explicit @xtype foo is converted to xtypes: {widget: ["foo"]}

The choice of names is currently misfortunate.  It would be logical
to use "alias", but we already have @alias which does a completely
different thing.  This issue has been disturbing us for long, but
now it really starts to hurt.  I feel we have to reclame @alias
from its current occupating and use some other tag like @include
to replace @alias.