Treat all package names equally when rendering tree.
This is a switch away from ext-doc style tree-logic where the tree
could look like this:

    - Foo
      - Bar
      - Bar.Baz
      - Bar.Zap

Now the tree always looks like this:

    - Foo
      - Bar
        - Baz
        - Zap

This change is mostly in response to request of a user working with
a system where most package names were also classes.  With previous
PackageLogic implementation he ended up with a pretty flat tree.