Show list of all source files of class.
Instead of :filename and :linenr properties being tied to class hash
itself, they are now inside :files property, which holds an array of
hashes which contain :filename and :linenr.

List of files is shown in the hierarchy sidebar. Especially useful for
classes like Ext.Element which are made up from many-many files.