Swap miniphone landscape and portrait styles.
The default 'landscape' now default to a view where the phone is in
portrait orientation, but because the screen is cut in half, the
resolution is actually a landscape.

Most importantly it allows us to simply write @example miniphone,
instead of having to write @example miniphone portrait - the default
landscape orientation is now the thing that one needs 99% of time.