Fix no-scrollbar bug in class tree.
The happened when the welcome page was first loaded and then one
switched to API tab. It happened because the class tree was activated
while it was hidden, so height calculations couldn't be performed
correctly. Fixed by adding an empty item to trees container that's
activated first instead of the class tree.