Embed support of JSB files into JSDuck itself.
At the moment only treating file as JSB if it has .jsb3 extension.
I'm not sure if the format of .jsb2 or whatever might be different,
so better be safe than sorry.

At the moment this is a bit of a hidden feature, but it could be
explained in manual.