Better automatic linking of class names.
Previous implementation was very hackish. Now only doing auto-detection
when not inside HTML tag (i.e. between <...>) and when not already
inside HTML link (i.e. between <a>...</a>).

As a result the #replace_class_names method is quite a bit simpler, as
it can be sure that it's always executed in valid context.