Speed up tests involving meta-tags.
The loading of all meta-tags at the start of several test-suites
slowed things down quite a bit.  Now only loading support for
@static in most places and only loading the full set of meta-tags
in one suite that tests all of them.

Also moving @markdown tag test to be together with other meta-tag tests.

Got the running time of tests down from 1.18 to 0.85 seconds.
That's about 1/4 faster. Not much, but it's a good goal to keep
the running time under a second.