Format every meta-tags only once.
The rendering of meta-tags is now performed inside ClassFormatter
like the rendering of most other stuff.  The result of rendering
is stored to :html_meta field (to keep the plain :meta field for
purposes of export).

The performance gains of doing it all once are very small, but
we should now get less duplicate error messages for bad @links,
which is nice.

Additionally all Exporter classes can now get rid of opts parameter,
which was only needed for instanciating DocFormatter, but that's
now done inside ClassFormatter.