Show info about overridden members.
For each member we now show a list of all the members that it overrides
(if there are any). No more needs one wonder if the method happens to
override a method in some parent class or mixin - the information is
right the in the docs.

This implementation mostly works. However I have to refactor it around a
bit to ensure it doesn't misbehave when parallel-processing is used.