Better display of mixins, requires, uses info.
Mixins are now separated to two groups: Direct mixins from current
class, Inherited mixins from parent classes.

Uses and Requires now also print warnings when unknown classes
referenced.

When any of these classes is missing, the class name will still be rendered
in overview page, but it won't be a link.

Additionally allow wildcard '*' in @requires/uses - don't throw warning
on it, but simply leave as is (not expanding the wildcard either).