Add --examples-base-url option.
This replaces the previous Docs.exampleBaseUrl JS-side option, which
was a bit of a hack.

Now all URL-s are prefixed in parsing stage, so the docs app doesn't need
to deal with the prefixing at all.

Additionally the externalUrl field in examples.json is no more supported.
Simply using the url field beginning with http:// will do the trick.