Order all static/instance methods in hover menus.
Previously statics were at the end, which made it seemingly look like
the menu wasn't correctly sorted.

The good thing was that the menu ordering matched the page ordering,
but whether it really mattered, remains to be seen when people start
to complain about this change.

For now I'm opting for the simpler code and uniformity of ordering
required/optional and instance/static.