Fix scrolling to hidden members.
When search directs to specific member that happens to be hidden
we now first show and expand it, and only after that scroll to it.
Expanding it first allows us to scroll it more correctly especially
in case when dealing with members at the very end of the page that
would otherwise be halfway hidden at the bottom.