Add [] lookup method to examples/guides/videos.
For this create a base class GroupedAsset that deals with
building the map of names.  To make the base class work, all
classes now use @groups instead of @examples/@videos/@guides
so they can be treated the same in GroupedAsset class.

The base class also has #each_item method to more easily iterate
over items in all groups.

For examples we are now also creating the "name" field into all
of its items.