Use example names in URL-s.
At the moment this changes nothing, as currently examples have no
names assigned and so the "url" field will be assigned to "name" field.
But it opens up the opportunity of assigning names to examples if
needed.

More importantly it makes the handling of examples more similar to
guides and videos.