Don't output original guides/videos/examples JSON files.
There needs to be a separate mechanism for giving people access to
all the assets used in documentation.  The previous approach of
writing out guides/videos/examples JSON files generated needless
files for most users of the docs, while not really helping the
people who wanted the actual source files of the docs, as these
files weren't really the actual source files, only gone-through-jsduck
versions of only some of them.