Use --examples-base-url also used for icons.
Instead of hardcoding ExtJS and Touch examples paths, the
--examples-base-url option determines where icons are located.

The important change is that examples.json file now has to define
icons relative to --examples-base-url (this was the case for Touch,
but not for ExtJS).

Example title and description fields are now also correctly convered,
independent of url being absolute or not.

Auto-created example name no more contains the --examples-base-url
in it.