Remove duplicated logic for showing notifications.
I didn't notice before that there was a #showError method already, so I
created #notify method to notify of subscription changes.  Merging these
two methods now into #notify to cover both uses.

The method now takes a third argument to allow for anchoring on
different sides.

One downside is, that I threw out the smooth fade-in and fade-out of
subscription notification.  It was a bit of a too much trouble for such
a small feature, so I guess the overall simplicity in code is a win.