Force all files to be read as UTF-8 in Ruby 1.9.
In Ruby 1.9 reading files through builtin IO class attempted to
auto-detect the encoding which sometimes failed.

Created JsDuck::IO to be used instead of it, which forces all files
to be treated as UTF-8.  It also behaves well in Ruby 1.8 - just
doing nothing with the input and causing no errors.