Support for optional parameter type definition.
To support Google Closure Compiler syntax for optional parameters.

    @param {Type=} name

This is detected at DocParser not in TypeParser like the rest of the
type definition syntax - all other way of defining optional parameters
are also implemented in DocParser, so it keeps things simpler to have it
all in there at one spot.