Fix Tab closing issue in IE8.
The code relied on the order of events being fired.  It expected that
when clicking tab close button the "click" handler for close button
will be invoked before the "click" handler for the whole tab.  This
worked in the majority of browsers but not in IE8 where the events
fire in the opposite order.

Now checking the event target inside tab click handler to see if
close button was clicked instead of relying on the close button handler
to set a flag.

This also fixes the issue of tabs sometimes not activating when clicked
in IE8.  Was caused by justClosed flag being set but not unset.