Fix class overview toolbar for ExtJS 4.1.
First I'm constraining the styles of the main class-overview container
inside the body-area - otherwise the styles will also have unpredictable
effects on toolbar (especially the styles for tables as in 4.1 the
search field inside toolbar is wrapped in two tables).

Secondly the search field styles are applied differently as the DOM that
makes up the field has changed considerably.  At the moment this results
in the field looking a bit different in 4.1.  Needs more work...

Also added width to the trigger-field which otherwise would change its
width when trigger is shown/hidden.