Avoid conflict with CodeMirror styles for <pre>.
CodeMirror just defines the rule:

    .CodeMirror pre { ... }

This selector has quite low specificity, so it's easy to accidentally
override it.  And this indeed happened when the selector for normal pre
elements was changed from:

    .class-overview pre { ... }

to:

    .class-overview .x-panel-body pre { ... }

The result was that a <pre> inside CodeMirror editor received unwanted
rounded corners etc, rendering the editor pretty-much unusable.

Resolved this by only styling pre.prettyprint and pre.notpretty.
(The latter class is now added to all not-to-be-prettified pre-s, which
distinguishes them from pre-s used inside CodeMirror component.)