Split EsprimaParser into Parser and Core parts.
Only the EsprimaCore is now singleton which runs esprima.js, the
EsprimaParser just uses the core and then does additional processing,
but most importantly EsprimaParser now uses the same interface as
JsParser - both of which are instanciated with source and then have
parse method invoked.  This allows for better handling of instance
variables and avoids possible conflicts when we do parallel processing.