Ignore the order of mixins in tests.
When auto-detecting mixins from code that has:

    mixins: {
        foo: "Some.Class",
	bar: "Other.Class"
    }

we are converting the JavaScript object to Ruby hash - but that means
the order becomes non-deterministing.  This resulted in tests checking
the list of mixins to be flaky as they expected a certain order.  But we
really don't care about the order, so I convert arrays to Sets and compere
these.