Restructure docset merging.
Merger#merge now takes the full docset as single parameter.  But the
docsets get pre-processed in SourceFile where the tagname of each of
them gets detected and each class-docset gets expanded into multiple
ones (either because the doc-comment contained configs/constructor, or
because class members were auto-detected from code).

Therefore the merger is now much simpler, having not to deal with the
complexities of @class docset.