Fix inline examples in IE8.
Using Ext.Element#on to attach load event handler to iframe element.
Assigning handler directly to iframe.onload didn't work in IE8.

I'm sure it also fixes the same problem in IE7 (at least it does so in
IE7-mode of IE8), and the fix itself is just applying the right way of
doing things.