Replace default value parsing with simpler implementation.
Instead of using JsLiteralParser, we just scan the string and attempt to
balance braces.  It's not really fool-proof, but should easily cover
more than 99% of use cases.  The only problem is when a default value
contains a string with "[" or "]" in it - but then it can be just
documented manually.

Getting rid of this use of JsLiteralParser and JsLiteralBuilder will
allow us to get rid of those classes together with JsParser completely -
replacing them with EsprimaJS.