Fix regression in sorting algorithm.
The sort data had been changed so that it now contained meta field
instead of private and removed fields directly.

Fixed it, so that public members are now again listed before deprecated,
private and removed ones (as a new thing now also taking account the
deprecated ones).

Enhanced the documentation to further explain how the algorithm works.

Eliminated lots of magic numbers from code and replaced them with
variables - most of which are calculated from the three main constants
defined at the beginning.  This should make it much simpler to enhance
the function in future.