Turn ParallelWrap into singleton class.
Otherwise we need to pass around the ParallelWrap instance or config
for it, which is annoying.  Parallelization is more of an implementation
detail - passing it explicitly to every class that needs to make use
of it doesn't really fit to the interface of these classes.  We
shouldn't need to change the interface when switching from normal each
loop to parallel each.  So a singleton class it is.