Throw away Gemfile, drop use of bundler.
Using both Gemfile for bundler and a .gemspec for gem doesn't seem to
work quite well together.  Besides I'm not really making much use of
bunder, so I feel I'm better off without this additional complexity.

Though, this means I need to explicitly list the dependencies to be
installed for development, but there are only a few, so it really
shouldn't be much of an issue.