Change how class members are internally stored.
Instead of storing members in two hashes - :members & :statics -
the members are now stored in simple :members array.

Lots of refactoring, but in the end the unit tests are all passing
and running JSDuck against ExtJS SDK works fine.  More work is needed
though as now JSDuck::Class contains several methods and algorithms
that are inappropriate in the context of new data structure.

Here I'm only changing the internals - export format is still the same
and so is the data structure given to generated Docs app.