Ensure users table entry when logging in.
Created Users class, which uses the ForumAuth to first log in through
Sencha Forum and then when needed creates entry to users table in
comments database.

The main thing is that now the user object stored into session contains
an `id` field which we can use in further queries to add/update/delete
comments.