Support nested grouping of assets.
Nested groups can be defined as follows:

    [{
        "title": "Main group",
        "items": [
            {"title: "subgroup",
             "items: [ ... ]}
        ]
    }]

In UI the tree will then show subfolders for the nested asset groups.
But the ThumbList display will still show all the assets in old way,
and the assets in subgroups will be expanded so they will appear as if
they were listed directly inside the top level group.

I guess in some way this beats the whole point of having the subgroups -
as all the assets still show up on the main page as if there were no
subgroups at all.  On the other hand we now technically have support for
subgroups, but I feel that's not really what these users were after when
they asked for nested grouping.

Additionally eliminated :dup_asset warning.  It didn't fit nicely to the
new way of iterating over assets, and I think it wasn't such a useful
warning anyway.