Extract BatchFormatter class from App.
This new class encapsulates the formatting of all classes.

Also, while formatting, we now create a new class_formatter for each
class that we process.  That's because otherwise the list of images
detected in one class will be added to the images detected from the next
one, and so on and on...  This wasn't really a problem because we
stripped away the duplicate images anyway, but it wasn't really the
right way to do it.