Eliminate recreation of database from testing.
Instead of dropping and recreating the database we now drop all the
tables and then recreate them.  This is because in our staging server I
currently have just one database available and I can't drop and recreate
it to run tests in it.

Additionally eliminated the update_votes.sql script - the triggers
themselves take care of votes being calculated.