New recent comments page.
The page is now also visible to normal users, with the only difference
that hideRead checkbox is only shown for moderators.

The page is divided into two sections: comments & users.

The comments section has two tabs which determine how the comments are
sorted (instead of a checkbox determining it).

The users section has a list of all users sorted by votes.