Filtering of users and topics lists.
Added field to the top of both users and topics which allows filtering
the lists below them.

Every time the filtering happens, the previously existing selection is
cleared.  This is maybe not the best behavior, but that's the compromise
I arrived at to prevent several edge-case bugs.  The problem is, that
when you filter out a currently selected record, it becomes deselected,
but the deselect event doesn't fire to let us know...