Init comments before launching Docs app.
Change Docs.Application to initialize comments before anything else.
This way the rest of the app can rely on comment counts to be
already loaded and avoid setting up listeners all over the place.

Additionally the CommentCounts is now accessed through Docs.Comments
and never directly as it's no more a singleton.