Sort mixins/requires/uses/etc in export.
This ensures that when running the export repeatedly the order
of mixins/requires/etc will stay the same.  Especially in the
case of mixins which are often defined in an object literal which
gets turned into a Ruby hash that messes up the order quite a bit.

By doing the sorting in Export::Full the Renderer no more needs
to sort these things.