Unify the responses of /users /targets /tags requests.
This change breaks compatibility with comments API of previous
JSDuck release.  But the only effect of this is that the Users and
Targets lists on comments index page will be empty, so I consider
it OK to make this change as it will only make the comments index
page a bit less useful while the transition happens, but won't
break anything badly.