Support for @class doc-comment right before object literal.
This allows for a way to document classes that aren't created with
a standard Ext.define:

    createClass("Blah", /** @class Blah */ {
        foo: 17,
        doSometing: function() {},
        doOtherThing: function() {}
    });

All the members inside the object literal will end up inside the class.

Additionally this is equivalent to the functionality of @lends tag in
jsdoc-toolkit, but we achieve the same without having a separate tag.