Improve speed of loading class docs into view.
Injecting of comment expanders is now deferred until member doc is
expanded.  This considerably cuts down the time of initial loading of
class docs.  Hovewer, clicking the expand-all button is now
really-really slow as clicking it triggers the injection of all those
expanders.  But I think that's a good trade-off for now.