Switch back to therubyracer 0.10.
Let's stay on the 0.10.x until the problems with 0.11.x get resolved.
This time though, don't specify exactly 0.10.1 but a range of versions.

With the 0.11.x therubyracer is attempting to use the system libv8,
and you have to install the libv8 gem by yourself when the system doesn't
have the libv8 available.  Additionally the compilation of libv8 is
quite time-taking and on some systems (like those with old Python)
will fail.  So lets hope these things get sorted out.

Although that somewhat complicates the install on Windows, as for it
only a 0.11.x version is available - so I need to generate a special
gem depending on 0.11.x from which I'll then build the executable.