Turn @private tag from MetaTag to Builtin.
Builtins::Tag now supports the @signature attribute like MetaTag.
For now the signature logic is duplicated as both tag groups support it.

In search data and class data export the data for builtin tags is
added to :meta field, so the frontend logic can stay the same.

On the good side, @private no more needs to be duplicated in both
member[:private] and member[:meta][:private] - just the former will do.