Make most StandardTagParser methods private.
Let the #hw method return it's original return value, which is the
DocScanner class - then it chained with one of the DocScanner
methods like #look or #match.  The only exception being #default_value
method which belongs to StandardTagParser and can't be therefore
chained from #hw method - but that's just a one little exception
we can live with.