Change Tag#process_doc API to take two params.
Instead of returning a hash that gets merged into the final class/member
hash, the latter hash gets passed in as the first parameter and
the #process_doc can add desired fields to it.

This gives #process_doc the ability to check what's already inside the
final hash and then take action depending on that - not just blindly
writing over anything that happens to be there.