Show "Nothing found" text when search fails.
That's much better than the previous flickering where before performing
the search the dropdown was shown and after getting no results it was
hidden again.

Better to just show a message "Nothing found" - this will also make
clear to the user that the search found no results and didn't just fail
for whatever reason.