Create MiniParser helper for use in parsing tests.
This helps to keep the parsing setup in one place and allow easier
changes to it when needed.  Each spec file only needs to call the
Helper::MiniParser.parse with a few arguments, simplifying the
setup code inside actual spec files.

Because the MiniParser.parse always returns Relations object, add
a #length method to Relations class, so that the tests that previously
checked the length of simple classes hash can now call the same method
on Relations object and continue passing.