Fix regex for detecting indented code.
For some reason I had always thought \Z is the end of string in Ruby,
but it turns out \z is actually the end of string while \Z will match
right before a newline at the end of string if there happens to be one.

This little difference caused quite an unexpected behavior in one corner
case where the code example is on just a line before the @tag.