Don't skip type parsing when doing export.
Why are we skipping it on that case?  Possibly to avoid creating
the :html_type field, but currently we are creating it anyway, which
defeats the whole purpose of skipping the parsing, plus we lose warnings
of bad typedefs.

Now we get the type warnings also when doing export. And the code
is simpler. Win-win.