Move creation of :files field from File to Merger.
This way the Tag#merge method will have access to both the filename and linenr,
which is useful for error reporting.

Previously the Merger introduced :linenr field, which was largely unused, except
for some tests checking it.  This resulted in exported JSON containing both :linenr
and :files fields - but the latter already has :linenr inside it.  So another good
step towards more consistency.