Eliminate the complex parameter merging logic.
Stick with the simple: if @param in comment, use that, otherwise
use params auto-detected from code.  This means the auto-merging
takes completely care of that.  So the merge method only needs to
set the defaults and report warnings.

This change eliminates support for the syntax where one defines
variabled types in doc-comment and names in code.  But that's a
very silly syntax anyway as one can't use it when one also wants
to document the variable.  So it's now gone and we have simplified
the code considerably.