Eliminate Tag::Autodetected class.
Handle the :autodetected field entirely inside Merger class.

From code we get the :autodetected hash with :tagname value - we
use this as the initial value for :autodetected field in merge result.
By adding this, we eliminate the need for the special merge logic
in Autodetected class (which turned out to be buggy anyway).