Exclude :inheritdoc and :localdoc from accessors.
The docs for accessors get auto-generated and we need nothing extra.
Having the :inheritdoc in there also causes errors when looking up
the parent.  For example the following:

    /**
     * @cfg {Number} width
     * @inheritdoc SomeClass#width
     * @accessor
     */

When the accessors would get the @inheritdoc, this would look like
we had the following methods:

    /**
     * @method getWidth
     * @inheritdoc SomeClass#width
     */

    /**
     * @method setWidth
     * @inheritdoc SomeClass#width
     */

But the inheritance system would attempt to inherit the docs from
"width" method, but there is no such thing because we were actually
inheriting from "width" config.

Therefore we're better off without the whole @inheritdoc in accessors.