Integrate guides full text search.
Initial version of the UI - the search results are combined with
API search and shown in the same old dropdown.  This means there
isn't quite enough room to show the full context of full text search
matches and the search is also slower, but it fits nicely to the
current UI.  Plus it was easy to implement.

The most questionable part is how the guide results are placed inside
API search results.  I'm drawing some arbitrary lines between the
scores given by guides search to split the results into boxes like
the API search does.