Invalidate all member indexes after InheritDoc processing.
After Process::InheritDoc has run we regenerate the ID-s of all
members and invalidate all member indexes.  The old code tried
to be clever and only recalculate members that had changed and
classes containing these, but it failed to do so correctly,
resulting in this hard-to-find bug.  Now I've taken a slower
approach of just recalculating everything, which should now
give us correct results.

Fixes #345